push

+++ lib/screens/note_list_screen.dart
@@ -68,13 +68,16 @@ class NoteList extends StatelessWidget {
                         document.imageUrl != null &&
                                 Uri.parse(document.imageUrl!).isAbsolute
                             ? ClipRRect(
-                                borderRadius: BorderRadius.circular(16),
+                                borderRadius: const BorderRadius.only(
+                                  topLeft: Radius.circular(16),
+                                  topRight: Radius.circular(16),
+                                ),
                                 child: Image.network(
                                   document.imageUrl!,
                                   fit: BoxFit.cover,
                                   alignment: Alignment.center,
                                   width: double.infinity,
-                                  height: 200,
+                                  height: 150,
                                 ),
                               )
                             : Container(),
@@ -83,7 +86,32 @@ class NoteList extends StatelessWidget {
                           subtitle: Text(document.description),
                           trailing: InkWell(
                             onTap: () {
-                              NoteService.deleteNote(document);
+                              showDialog(
+                                context: context,
+                                builder: (BuildContext context) {
+                                  return AlertDialog(
+                                    title: const Text('Konfirmasi Hapus'),
+                                    content: Text(
+                                        'Yakin ingin menghapus data \'${document.title}\' ?'),
+                                    actions: <Widget>[
+                                      TextButton(
+                                        child: const Text('Cancel'),
+                                        onPressed: () {
+                                          Navigator.of(context).pop();
+                                        },
+                                      ),
+                                      TextButton(
+                                        child: const Text('Hapus'),
+                                        onPressed: () {
+                                          NoteService.deleteNote(document)
+                                              .whenComplete(() =>
+                                                  Navigator.of(context).pop());
+                                        },
+                                      ),
+                                    ],
+                                  );
+                                },
+                              );
                             },
                             child: const Padding(
                               padding: EdgeInsets.symmetric(vertical: 10),
